JsonNet MediaTypeHandler updates
