--- src/Simple.Web.JsonNet.Tests/HalReadTests.cs
@@ -1,44 +1,0 @@
-﻿namespace Simple.Web.JsonNet.Tests
-{
-    using System.Text;
-    using System.IO;
-    using JsonNet;
-    using TestHelpers.Sample;
-    using Xunit;
-
-    public class HalReadTests
-    {
-        [Fact]
-        public void ReadsBasicObject()
-        {
-            const string source = @"{""name"":""Arthur Dent"",""location"":""Guildford""}";
-            Person actual;
-            using (var stream = new MemoryStream(Encoding.UTF8.GetBytes(source)))
-            {
-                stream.Position = 0;
-                var target = new HalJsonMediaTypeHandler();
-                actual = target.Read<Person>(stream).Result;
-            }
-
-            Assert.Equal("Arthur Dent", actual.Name);
-            Assert.Equal("Guildford", actual.Location);
-        }
-
-        [Fact]
-        public void ReadsObjectWithLinks()
-        {
-            const string source =
-                @"{""_links"": {""self"":""/person/42""}, ""name"":""Arthur Dent"",""location"":""Guildford""}";
-            Person actual;
-            using (var stream = new MemoryStream(Encoding.UTF8.GetBytes(source)))
-            {
-                stream.Position = 0;
-                var target = new HalJsonMediaTypeHandler();
-                actual = target.Read<Person>(stream).Result;
-            }
-
-            Assert.Equal("Arthur Dent", actual.Name);
-            Assert.Equal("Guildford", actual.Location);
-        }
-    }
-}
